cd/world/field: drop redundant externs, fix prototype mismatches
The previous round of header consolidation surfaced four conflicts
when 'make full' rebuilt every TU from scratch:

  cd.h::func_80038868: prototype 3rd-arg type was wrong (had 'u32 dest',
    src/cdrom.c defines it as 'u8 *dest'). Fix the header.

  cd.h::func_800393C8: the original cdrom.c had a local 'extern s32
    func_800393C8(void)' that callers across the codebase depend on
    (they consume $v0 as a 32-bit value, not the u8 the implementation
    returns). cdread.c had it typed as 'u8' which produced an 'andi
    v0, v0, 0xff' mask at every caller and grew cdReadAsyncSync,
    loadFileTable, initSoundEngine, ... by 4-12 bytes each. Switch the
    cdread.c definition to 's32' (the body still returns the byte field
    D_8008A3D8.status, just zero-extended).

  cdrom.c: drop the local 'extern s32 func_800393C8(void)' — comes
    from <cd.h> now.

  world/we_object2.c: drop the local 'extern s32 sndCmd10', 'extern
    void sndCmd11', 'extern void sndCmdC1' — sound.h now declares
    them all with consistent 's32' return types matching the asm.

  field/fe_object4.c: drop the local 'extern s32 D_8008538C' — field.h
    now declares it as 'Eline *' to match the actual semantics.

+++ src/cdread.c
@@ -325,7 +325,7 @@ void cdBreakRead(void) {
  *
  * @return Current CD read status after handler execution.
  */
-u8 func_800393C8(void) {
+s32 func_800393C8(void) {
     D_800562D8[D_8008A3D8.status]();
     return D_8008A3D8.status;
 }
